Bug 1016423 - Use Hawk for making authenticated HTTP requests and support FxA assertions

+++ js/client_request_helper.js
@@ -1,13 +1,18 @@
 /* -*- Mode: js; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- /
 /* vim: set shiftwidth=2 tabstop=2 autoindent cindent expandtab: */
 
+/* exported ClientRequestHelper */
+
+/* globals Config */
+
 'use strict';
 
 (function(exports) {
-  // TODO Poing to the right server and retrieve this info
-  // from a 'config' file
-  var SERVER_URL = 'http://loop.dev.mozaws.net';
+  var SERVER_URL = Config.server_url;
   var TIMEOUT = 15000;
+
+  /** HAWK credentials */
+  var _hawkCredentials = null;
 
   function _callback(cb, args) {
     if (cb && typeof cb === 'function') {
@@ -16,19 +21,36 @@
   }
 
   function _request(options, onsuccess, onerror) {
-    var req = new XMLHttpRequest();
+    var req = new XMLHttpRequest({mozSystem: true});
     req.open(options.method, options.url, true);
     req.setRequestHeader('Content-Type', 'application/json');
     req.responseType = 'json';
     req.timeout = TIMEOUT;
-    req.withCredentials = true;
 
+    var authorization = '';
+    if (options.credentials) {
+      switch (options.credentials.type) {
+        case 'BrowserID':
+          authorization =
+            options.credentials.type + ' ' + options.credentials.value;
+          break;
+        default:
+          var hawkHeader = hawk.client.header(options.url, options.method, {
+            credentials: options.credentials.value
+          });
+          authorization = hawkHeader.field;
+          break;
+      }
+      req.setRequestHeader('authorization', authorization);
+    }
     req.onload = function() {
       if (req.status !== 200 && req.status !== 302) {
-        _callback(onerror, [req.response]);
+        _callback(onerror, [req.statusText]);
         return;
       }
-      _callback(onsuccess, [req.response]);
+      _callback(
+        onsuccess, [req.response, req.getResponseHeader('Hawk-Session-Token')]
+      );
     };
 
     req.onerror = req.ontimeout = function(event) {
@@ -48,24 +70,65 @@
       return SERVER_URL;
     },
 
-    register: function register(pushEndpoint, onsuccess, onerror) {
+    signUp: function signUp(credentials, pushEndpoint, onsuccess, onerror) {
       _request({
-        method: 'POST',
-        url: SERVER_URL + '/registration',
-        body: {
-          simple_push_url: pushEndpoint
-        }
-      }, onsuccess, onerror);
+          method: 'POST',
+          url: SERVER_URL + '/registration',
+          body: {
+            simple_push_url: pushEndpoint
+          },
+          credentials: credentials
+        },
+        function onSuccess(result, sessionToken) {
+          if (!sessionToken) {
+            _callback(onerror, [new Error('No session token')]);
+            return;
+          }
+          deriveHawkCredentials(sessionToken, 'sessionToken', 2 * 32,
+            function(hawkCredentials) {
+              _hawkCredentials = {
+                type: 'Hawk',
+                value: hawkCredentials
+              };
+              _callback(onsuccess, [result, _hawkCredentials]);
+          });
+        },
+        onerror);
+    },
+
+    signIn: function signIn(credentials, pushEndpoint, onsuccess, onerror) {
+      _request({
+          method: 'POST',
+          url: SERVER_URL + '/registration',
+          body: {
+            simple_push_url: pushEndpoint
+          },
+          credentials: credentials
+        },
+        function onSuccess(result) {
+          _hawkCredentials = credentials;
+          _callback(onsuccess, [result]);
+        },
+        onerror
+      );
     },
 
     generateCallUrl: function generateCallUrl(callerId, onsuccess, onerror) {
+      if (!_hawkCredentials) {
+        _callback(onerror, [new Error('No HAWK credentials')]);
+        return;
+      }
       _request({
-        method: 'POST',
-        url: SERVER_URL + '/call-url',
-        body: {
-          callerId: callerId
-        }
-      }, onsuccess, onerror);
+          method: 'POST',
+          url: SERVER_URL + '/call-url',
+          body: {
+           callerId: callerId
+          },
+          credentials: _hawkCredentials
+        },
+        onsuccess,
+        onerror
+      );
     },
 
     getCallUrl: function getCallUrl(token, onsuccess, onerror) {
@@ -83,17 +146,33 @@
     },
 
     deleteCall: function deleteCall(token, onsuccess, onerror) {
+      if (!_hawkCredentials) {
+        _callback(onerror, [new Error('No HAWK credentials')]);
+        return;
+      }
       _request({
-        method: 'DELETE',
-        url: SERVER_URL + '/calls/' + token
-      }, onsuccess, onerror);
+          method: 'DELETE',
+          url: SERVER_URL + '/calls/' + token,
+          credentials: _hawkCredentials
+        },
+        onsuccess,
+        onerror
+      );
     },
 
     getCalls: function getCalls(version, onsuccess, onerror) {
+      if (!_hawkCredentials) {
+        _callback(onerror, [new Error('No HAWK credentials')]);
+        return;
+      }
       _request({
-        method: 'GET',
-        url: SERVER_URL + '/calls?version=' + version
-      }, onsuccess, onerror);
+          method: 'GET',
+          url: SERVER_URL + '/calls?version=' + version,
+          credentials: _hawkCredentials
+        },
+        onsuccess,
+        onerror
+      );
     },
 
     getCall: function getCall(callId, onsuccess, onerror) {
